feat: implement standardized JSON responses across controllers; add middleware for token type and verified email checks; update policies for role-based access control; refactor API routes for improved organization and security

+++ routes/api.php
@@ -16,8 +16,8 @@ Route::get('/', function () {
 });
 
 Route::prefix('auth')->controller(AuthController::class)->group(function () {
-    Route::post('web-login', 'web_login');
-    Route::post('mobile-login', 'mobile_login');
+    Route::post('web-login', 'web_login')->middleware('tokenType:web');
+    Route::post('mobile-login', 'mobile_login')->middleware('tokenType:mobile');
     Route::post('register', 'register');
 });
 
@@ -25,52 +25,59 @@ Route::middleware(['auth:sanctum'])->group(function () {
 
     // Posts
     Route::prefix('posts')->controller(PostController::class)->group(function () {
-        Route::get('deleted', 'deleted')->middleware('hasRoles:admin');
-        Route::get('restore/{id}', 'restore');
-        Route::delete('force-delete/{id}', 'force_delete')->middleware('hasRoles:admin', 'isActive');
+        Route::get('deleted', 'deleted')->middleware(['hasRoles:admin,manager']);
+        Route::get('restore/{id}', 'restore')->middleware(['hasRoles:admin']);
+        Route::delete('force-delete/{id}', 'force_delete')->middleware(['hasRoles:admin', 'isActive']);
     });
 
     // Comments
-    Route::prefix('comments')->controller(CommentController::class)->group(function () {
-        Route::get('deleted', 'deleted');
-        Route::get('restore/{id}', 'restore');
-        Route::delete('force-delete/{id}', 'force_delete');
+    Route::prefix('comments')->controller(CommentController::class)->middleware(['isActive', 'verifiedEmail'])->group(function () {
+        Route::get('deleted', 'deleted')->middleware('hasRoles:admin,manager');
+        Route::get('restore/{id}', 'restore')->middleware('hasRoles:admin');
+        Route::delete('force-delete/{id}', 'force_delete')->middleware('hasRoles:admin');
     });
 
     // Replies
-    Route::prefix('replies')->controller(ReplyController::class)->group(function () {
-        Route::get('deleted', 'deleted');
-        Route::get('restore/{id}', 'restore');
-        Route::delete('force-delete/{id}', 'force_delete');
+    Route::prefix('replies')->controller(ReplyController::class)->middleware(['isActive', 'verifiedEmail'])->group(function () {
+        Route::get('deleted', 'deleted')->middleware('hasRoles:admin,manager');
+        Route::get('restore/{id}', 'restore')->middleware('hasRoles:admin');
+        Route::delete('force-delete/{id}', 'force_delete')->middleware('hasRoles:admin');
+    });
+
+    // Reactions
+    Route::prefix('reactions')->controller(ReactionController::class)->middleware(['isActive', 'verifiedEmail'])->group(function () {
+        Route::get('deleted', 'deleted')->middleware('hasRoles:admin,manager');
+        Route::get('restore/{id}', 'restore')->middleware('hasRoles:admin');
+        Route::delete('force-delete/{id}', 'force_delete')->middleware('hasRoles:admin');
+    });
+
+    // Post-Statuses
+    Route::prefix('post-statuses')->controller(PostStatusController::class)->middleware(['isActive', 'verifiedEmail'])->group(function () {
+        Route::get('deleted', 'deleted')->middleware('hasRoles:admin,manager');
+        Route::get('restore/{id}', 'restore')->middleware('hasRoles:admin');
+        Route::delete('force-delete/{id}', 'force_delete')->middleware('hasRoles:admin');
+    });
+
+    // Reaction-Types
+    Route::prefix('reaction-types')->controller(ReactionTypeController::class)->middleware(['isActive', 'verifiedEmail'])->group(function () {
+        Route::get('deleted', 'deleted')->middleware('hasRoles:admin,manager');
+        Route::get('restore/{id}', 'restore')->middleware('hasRoles:admin');
+        Route::delete('force-delete/{id}', 'force_delete')->middleware('hasRoles:admin');
     });
 
     // Users
     Route::prefix('users')->controller(UserController::class)->group(function () {
-        Route::get('deleted', 'deleted');
-        Route::get('restore/{id}', 'restore');
-        Route::delete('force-delete/{id}', 'force_delete');
+        Route::get('deleted', 'deleted')->middleware('hasRoles:admin');
+        Route::get('restore/{id}', 'restore')->middleware('hasRoles:admin');
+        Route::delete('force-delete/{id}', 'force_delete')->middleware(['hasRoles:admin', 'isActive']);
     });
 
-    // Reactions
-    Route::prefix('reactions')->controller(ReactionController::class)->group(function () {
-        Route::get('deleted', 'deleted');
-        Route::get('restore/{id}', 'restore');
-        Route::delete('force-delete/{id}', 'force_delete');
-    });
-
-    // Post-Statuses
-    Route::prefix('post-statuses')->controller(PostStatusController::class)->group(function () {
-        Route::get('deleted', 'deleted');
-        Route::get('restore/{id}', 'restore');
-        Route::delete('force-delete/{id}', 'force_delete');
-    });
-
-    // Reaction-Types
-    Route::prefix('reaction-types')->controller(ReactionTypeController::class)->group(function () {
-        Route::get('deleted', 'deleted');
-        Route::get('restore/{id}', 'restore');
-        Route::delete('force-delete/{id}', 'force_delete');
-    });
+    // Dashboard
+    // Route::prefix('dashboard')->controller(DashboardController::class)->group(function () {
+    //     Route::get('deleted', 'deleted');
+    //     Route::get('restore/{id}', 'restore');
+    //     Route::delete('force-delete/{id}', 'force_delete');
+    // });
 
     // Posts, comments, reactions, users, replies
     Route::apiResources(
@@ -86,14 +93,14 @@ Route::middleware(['auth:sanctum'])->group(function () {
     );
 
     // Auth
-    Route::prefix('auth')->controller(AuthController::class)->group(function () {
+    Route::prefix('auth')->controller(AuthController::class)->middleware(['isActive', 'verifiedEmail'])->group(function () {
 
         // All Sessions
         Route::prefix('sessions')->group(function () {
             Route::get('active', 'active_sessions');
             Route::get('current', 'current_session');
             Route::get('others', 'other_sessions');
-            Route::get('{id}', 'show_session');
+            Route::get('{id}', 'show_session')->middleware('hasRoles:admin,manager');
         });
 
         // logout
@@ -101,7 +108,7 @@ Route::middleware(['auth:sanctum'])->group(function () {
             Route::post('all', 'logout_all');
             Route::post('current', 'logout_current');
             Route::post('others', 'logout_others');
-            Route::post('{id}', 'logout_session');
+            Route::get('{id}', 'logout_session')->middleware('hasRoles:admin,manager');
         });
 
         // Profile
